tests/rlcmac/ts_44_018_test: do not depend on libosmogsm
Header file <osmocom/gsm/protocol/gsm_04_08.h> belongs to libosmogsm.
However, file 'contrib/libosmo-gprs.spec.in' does not list this lib
as a build dependency.  So building the RPM package currently fails.

We could have listed libosmogsm as 'BuildRequires', but given that
the only thing we use from this lib is the GSM_MACBLOCK_LEN macro,
there is no real need to depend on it.

Change-Id: Ieac350e7a05965c64e8ccf8dd17a569f0afd9514