rlcmac: add decoder and test vectors for IA Rest Octets
This new API is needed for the MS side GPRS implementation in order
to be able to establish an Uplink TBF.  Both wireshark and osmo-pcu
rely on hand-written code for parsing and generating the IA Rest
Octets IE, making no use of the CSN.1 codec.

We already have some (mostly commented out) CSN.1 definitions needed
for decoding the IA Rest Octets IE, inherited as-is from wireshark.
Ths patch adds the missing bits and a few unit test vectors.

Note that decoding of the LH part, specifically the EGPRS Packet
Uplink Assignment structure, is currently done incorrectly because
there is a problem in the IA_EGPRS_t definition.  This will be
fixed in a follow-up patch.

For reference, see 3GPP TS 44.018, table 10.5.2.16.1.

Change-Id: I39a29dc9b5b22ce4374ae33336696014e326d012
Related: OS#5500