sndcp: Initial support for async SN-XID.ind and SN-XID.rsp
Before this patch, the response was being crafted synchronously through
a direct function call. This patch provides initial support to trigger
the primitives so that the application using the SNDCP layer can handle
the indication and submit a response.
The different XID L3 params are still not being passed to the app, that
will be done in a subsequent patch.

Change-Id: I0a5069fd3dc0d6c3dd28aeae09b51c49dd8be92d