gmm: Introduce interface GMMBSSGP
This interface was not described in TS 24.007 like the rest, though some
references to it can be seen in that document in Annex C.

The interface is actually described in TS 48.018 5.3.

This commit only adds the skeleton with all the primitives, etc.
All the logic inside GMM layer implementation, and primitive specific fields,
is still missing.

Related: OS#5852
Change-Id: I39045833fd43cfe98cb1a3812fbce3fdcaae6dc6